Extract workspace model context extension
