+++ Sources/QuillCodeApp/WorkspaceModel.swift
@@ -73,30 +73,6 @@ public final class QuillCodeWorkspaceModel {
         mcpRuntime.terminateAllRunningProcesses()
     }
 
-    public var selectedThread: ChatThread? {
-        guard let selectedThreadID = root.selectedThreadID else { return nil }
-        return root.threads.first { $0.id == selectedThreadID }
-    }
-
-    public var selectedProject: ProjectRef? {
-        guard let selectedProjectID = root.selectedProjectID else { return nil }
-        return root.projects.first { $0.id == selectedProjectID }
-    }
-
-    public var activeWorkspaceRoot: URL? {
-        guard let selectedProject, !selectedProject.isRemote else { return nil }
-        return URL(fileURLWithPath: selectedProject.path)
-    }
-
-    var terminalCurrentDirectoryURL: URL? {
-        WorkspaceTerminalEngine.currentDirectoryURL(
-            terminal: terminal,
-            selectedProjectID: knownProjectID(root.selectedProjectID),
-            selectedProjectIsRemote: selectedProject?.isRemote == true,
-            activeWorkspaceRoot: activeWorkspaceRoot
-        )
-    }
-
     func syncTerminalSessionToSelectedProject() {
         WorkspaceTerminalEngine.syncSessionToSelectedProject(
             terminal: &terminal,
@@ -109,29 +85,6 @@ public final class QuillCodeWorkspaceModel {
         _ mutation: (inout BrowserState, inout String?) -> Result
     ) -> Result {
         mutation(&browser, &lastError)
-    }
-
-    public var currentToolCards: [ToolCardState] {
-        guard let selectedThread else { return [] }
-        let cards = WorkspaceTranscriptSurfaceBuilder(thread: selectedThread).toolCards()
-        return executionContextSurfaceBuilder.enrichToolCards(cards, for: selectedThread)
-    }
-
-    public var currentTimelineItems: [TranscriptTimelineItemSurface] {
-        guard let selectedThread else { return [] }
-        let items = WorkspaceTranscriptSurfaceBuilder(thread: selectedThread).timelineItems()
-        return executionContextSurfaceBuilder.enrichTimelineItems(items, for: selectedThread)
-    }
-
-    private var executionContextSurfaceBuilder: WorkspaceExecutionContextSurfaceBuilder {
-        WorkspaceExecutionContextSurfaceBuilder(
-            selectedProject: selectedProject,
-            projects: root.projects
-        )
-    }
-
-    func project(id: UUID) -> ProjectRef? {
-        root.projects.first { $0.id == id }
     }
 
     func appendNotice(_ summary: String) {
