+++ Tests/QuillCodeParityTests/ParityWorkspaceModelGateTests.swift
@@ -129,13 +129,27 @@ final class ParityWorkspaceModelGateTests: QuillCodeParityTestCase {
 
     func testWorkspaceModelDelegatesExecutionContextSurfaceBuilding() throws {
         let modelText = try Self.appSourceText(named: "WorkspaceModel.swift")
+        let contextText = try Self.appSourceText(named: "WorkspaceModelContext.swift")
         let builderText = try Self.appSourceText(named: "WorkspaceExecutionContextSurfaceBuilder.swift")
 
+        XCTAssertTrue(contextText.contains("extension QuillCodeWorkspaceModel"), "Workspace context APIs should live in a focused model extension.")
+        XCTAssertTrue(contextText.contains("public var selectedThread"), "Selected thread lookup should live in the context extension.")
+        XCTAssertTrue(contextText.contains("public var selectedProject"), "Selected project lookup should live in the context extension.")
+        XCTAssertTrue(contextText.contains("public var activeWorkspaceRoot"), "Active local workspace lookup should live in the context extension.")
+        XCTAssertTrue(contextText.contains("var terminalCurrentDirectoryURL"), "Terminal current-directory lookup should live in the context extension.")
+        XCTAssertTrue(contextText.contains("public var currentToolCards"), "Current tool-card projection should live in the context extension.")
+        XCTAssertTrue(contextText.contains("public var currentTimelineItems"), "Current timeline projection should live in the context extension.")
         XCTAssertTrue(builderText.contains("struct WorkspaceExecutionContextSurfaceBuilder"), "Execution context enrichment should live in a focused builder.")
         XCTAssertTrue(builderText.contains("func enrichToolCards("), "Tool-card context enrichment should be directly testable.")
         XCTAssertTrue(builderText.contains("func enrichTimelineItems("), "Timeline context enrichment should be directly testable.")
         XCTAssertTrue(builderText.contains("static func isProjectExecutionTool"), "Project-execution tool classification should be directly testable.")
-        XCTAssertTrue(modelText.contains("WorkspaceExecutionContextSurfaceBuilder("), "WorkspaceModel should delegate execution-context enrichment.")
+        XCTAssertTrue(contextText.contains("WorkspaceExecutionContextSurfaceBuilder("), "WorkspaceModel context APIs should delegate execution-context enrichment.")
+        XCTAssertFalse(modelText.contains("public var selectedThread"), "WorkspaceModel.swift should not own selected thread lookup.")
+        XCTAssertFalse(modelText.contains("public var selectedProject"), "WorkspaceModel.swift should not own selected project lookup.")
+        XCTAssertFalse(modelText.contains("public var activeWorkspaceRoot"), "WorkspaceModel.swift should not own active workspace root lookup.")
+        XCTAssertFalse(modelText.contains("public var currentToolCards"), "WorkspaceModel.swift should not own current tool-card projection.")
+        XCTAssertFalse(modelText.contains("public var currentTimelineItems"), "WorkspaceModel.swift should not own current timeline projection.")
+        XCTAssertFalse(modelText.contains("WorkspaceExecutionContextSurfaceBuilder("), "WorkspaceModel.swift should not own execution-context enrichment.")
         XCTAssertFalse(modelText.contains("private func enrichToolCards"), "WorkspaceModel should not own tool-card context enrichment.")
         XCTAssertFalse(modelText.contains("private func enrichTimelineItems"), "WorkspaceModel should not own timeline context enrichment.")
         XCTAssertFalse(modelText.contains("private static func isProjectExecutionTool"), "WorkspaceModel should not own project-execution tool classification.")
