+++ docs/CODE_QUALITY_AUDIT.md
@@ -23,7 +23,7 @@ The architecture is moving in the right direction: core state is value typed, pe
 
 | File | Grade | Next Improvement |
 | --- | --- | --- |
-| `Sources/QuillCodeApp/WorkspaceModel.swift` | A- | Command parsing, automation records/run drafts, terminal session construction, project registry transitions, context/action lookup, review-comment planning, tool override composition, SSH Remote tool execution, browser location/state transitions, MCP surface state, MCP request parsing, MCP runtime/catalog/launch work, tool-card surface types, execution-context enrichment, thread seeding, thread lifecycle transitions, thread persistence, local command transcript mutation, thread notice mutation, sidebar selection transitions, sidebar bulk action planning, project context refresh, `/status` context assembly, and top-bar state assembly now live in focused helpers; keep extracting pure surface/workflow builders before adding more parity commands. |
+| `Sources/QuillCodeApp/WorkspaceModel.swift` | A- | Command parsing, automation records/run drafts, terminal session construction, project registry transitions, context/action lookup, review-comment planning, tool override composition, SSH Remote tool execution, browser location/state transitions, MCP surface state, MCP request parsing, MCP runtime/catalog/launch work, tool-card surface types, execution-context enrichment, thread seeding, thread lifecycle transitions, thread persistence, local command transcript mutation, thread notice mutation, sidebar selection transitions, sidebar bulk action planning, project context refresh, selected context queries, `/status` context assembly, and top-bar state assembly now live in focused helpers; keep extracting pure surface/workflow builders before adding more parity commands. |
 | `Sources/QuillCodeApp/WorkspaceSwiftUIView.swift` | A | The shell is now top-bar/sidebar chrome, state, and routing; center-pane layout, workspace sheet presentation, and worktree dialog lifecycle live in focused files. Keep future modal families and command workflow rules out of the root shell. |
 | `Sources/QuillCodeApp/QuillCodeWorkspaceMainPaneView.swift` | A- | Center-pane layout owns transcript/browser/extensions/memories/terminal/composer/activity composition and runtime issue recovery wiring. Keep workflow decisions in planners and avoid growing this into a second root shell. |
 | `Sources/QuillCodeApp/QuillCodeToolCardView.swift` | A- | Native tool-card composition is now separate from reusable controls, artifact previews, and raw detail blocks. Keep future status/action chrome in `QuillCodeToolCardControls.swift` and artifact rendering in `QuillCodeToolArtifactViews.swift`. |
@@ -117,6 +117,18 @@ Changes:
 - Added `QuillCodeDesktopCommandCoordinator` and `QuillCodeDesktopCommandPerforming` so typed command action dispatch has one focused owner.
 - Rewired `QuillCodeDesktopController` to plan commands, delegate action dispatch, and keep only concrete UI/workspace capabilities such as opening settings, toggling panes, and running workspace commands.
 - Extended the desktop parity gate so typed action switching stays in the command coordinator and cannot drift back into the controller.
+
+## 2026-06-25 Workspace Context Extension Pass
+
+Overall grade after this slice: **A read-side context ownership, A execution-context enrichment boundary, A- central model size**.
+
+`WorkspaceModel.swift` still owned selected thread/project lookup, active local workspace lookup, terminal current-directory lookup, and current transcript/tool-card projection queries. These are read-side context surfaces, not root storage or mutation policy, and they are used broadly across workflows.
+
+Changes:
+
+- Added `WorkspaceModelContext.swift` for selected thread/project lookup, active workspace root, terminal current directory, current tool cards, current timeline items, and project lookup.
+- Kept execution-context enrichment delegated to `WorkspaceExecutionContextSurfaceBuilder`.
+- Extended the workspace model parity gate so selected context queries and current transcript projections stay out of the root model file.
 
 ## 2026-06-25 Desktop Automation Coordinator Pass
