+++ docs/DECISIONS.md
@@ -332,6 +332,7 @@
 - Tool-run lifecycle status planning lives in `WorkspaceToolRunLifecyclePlanner`. `WorkspaceModel.runToolCall` still applies actor-isolated state, persistence, event recording, and browser/error mutation, but start error clearing and final idle/failed status selection should stay in the focused planner so command execution lifecycle rules remain directly testable.
 - Terminal run lifecycle status planning lives in `WorkspaceTerminalLifecyclePlanner`. `WorkspaceModel.runTerminalCommand` still owns async streaming, terminal entry mutation, session marker cleanup, and top-bar application, but started, missing-context, stopped, cancelled, and final idle/failed status selection should stay in the focused planner so terminal behavior stays aligned with tool-run lifecycle boundaries.
 - Active-work Stop All and Disconnect All lifecycle status planning lives in `WorkspaceActiveWorkStopPlanner`. `WorkspaceModel` still owns cancelling MCP servers, send state, terminal state, remote-project detachment, and top-bar application, but cancel/disconnect error clearing plus stopped/idle status selection should stay in the focused planner so toolbar commands follow the same testable lifecycle boundary as tool and terminal runs.
+- `WorkspaceModel.swift` owns stored app state, persistence handles, and shared mutation primitives. Read-side context queries live in `WorkspaceModelContext.swift`: selected thread/project, active workspace root, terminal current directory, current tool cards, current timeline items, and project lookup. Current transcript/tool-card projections still delegate execution-context enrichment to `WorkspaceExecutionContextSurfaceBuilder`.
 - Command routing must be explicit. `WorkspaceCommandRoutingCatalog` classifies host-owned commands such as Stop All and Computer Use permission actions separately from command IDs that `WorkspaceCommandPlan` can execute in the workspace model. SwiftUI and desktop planners reject unknown IDs instead of dispatching them into silent no-ops, and command-surface coverage proves every emitted command is either presentational or dispatchable.
 - Local git, git hunk, git worktree, and git router coverage lives in focused QuillCodeTools suites, not the mixed `ToolTests.swift` catch-all. `GitLocalToolExecutorTests`, `GitPatchToolExecutorTests`, `GitWorktreeToolExecutorTests`, and `GitToolRouterTests` share fixtures from `ToolTestSupport.swift` so git behavior can evolve without turning one broad file into a failure triage bottleneck.
 - `ToolTests.swift` is retired. File primitives, generic apply-patch primitives, shell router boundary checks, shell executor behavior, git tools, GitHub PR tools, and MCP probing each live in named focused suites, with `ParityToolGateTests` guarding the retirement of the broad catch-all.
